+++ docs/src/app/pages/docs/deck/app-deck-pager/app-deck-pager.tsx
@@ -34,25 +34,33 @@ export class AppDeckPager {
 <p><a href="https://deckdeckgo.com">DeckDeckGo</a> will per default display a pager in form of a progress circle bar. It&#39;s possible to hide it or to customize the following various style options.</p>
 <p>Also worth to notice that the pager inherits per default the document and deck direction (LTR or RTL for example).</p>
 <ul>
-<li><a href="#app-deck-pager-show-or-hide">Show or hide</a></li>
+<li><a href="#app-deck-pager-deck">Deck</a></li>
 <li><a href="#app-deck-pager-customization">Customization</a></li>
 </ul>
 <h2 id="app-deck-pager-show-or-hide">Show or hide</h2>
-<p>The show or hide options of the pager are available on the <code>&lt;deckgo-deck&gt;</code> element.</p>
+<p>To show or hide the pager, a CSS4 variable has to be set on the <code>&lt;deckgo-deck&gt;</code> element. Two more options are also available when set on this element.</p>
 <table>
 <thead>
 <tr>
 <th>Attribute</th>
-<th>Type</th>
 <th>Default</th>
 <th>Description</th>
 </tr>
 </thead>
 <tbody><tr>
-<td>pagerPercentage</td>
-<td>boolean</td>
-<td>true</td>
-<td>Show or hide the progression in percentage inside the pager</td>
+<td>--pager-display</td>
+<td></td>
+<td>The display property of the pager. Set to <code>none</code> to hide it.</td>
+</tr>
+<tr>
+<td>--pager-position-left</td>
+<td></td>
+<td>The left attribute of the absolute positioning of the pager over the deck</td>
+</tr>
+<tr>
+<td>--pager-position-right</td>
+<td></td>
+<td>The right attribute of the absolute positioning of the pager over the deck</td>
 </tr>
 </tbody></table>
 <h2 id="app-deck-pager-customization">Customization</h2>
@@ -116,21 +124,17 @@ export class AppDeckPager {
 <td></td>
 </tr>
 <tr>
-<td>--pager-position-left</td>
-<td></td>
-<td></td>
+<td>--pager-text-percentage-display</td>
+<td>hidden</td>
+<td>Set to <code>block</code> to display a progression with percentage (for example: 35%)</td>
 </tr>
 <tr>
-<td>--pager-position-right</td>
-<td></td>
-<td></td>
-</tr>
-<tr>
-<td>--pager-display</td>
-<td></td>
-<td>To hide the pager</td>
+<td>--pager-text-slides-display</td>
+<td>hidden</td>
+<td>Set to <code>block</code> to display a progression as slides&#39; count (for example: 2/15)</td>
 </tr>
 </tbody></table>
+<p>Note: of course if you would display both <code>--pager-text-percentage-display</code> and <code>--pager-text-slides-display</code> it would be a weird display, use just one at once.</p>
 </main>
 
         <app-footer></app-footer>